# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Admin API",
        "description": "API Documentation for the Laravel MongoDB MVC backend.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List courses",
                "operationId": "listCourses",
                "responses": {
                    "200": {
                        "description": "Course collection.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a course",
                "operationId": "createCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Learn the fundamentals of Laravel.",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": "665c9f2e8f1b2a4d7c123457",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/courses/{course}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a course",
                "operationId": "getCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "MongoDB course ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Course not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Replace a course",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "MongoDB course ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Advanced Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Build production-ready applications.",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": "665c9f2e8f1b2a4d7c123457",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Course not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "MongoDB course ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course deleted successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Course not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Courses"
                ],
                "summary": "Partially update a course",
                "operationId": "patchCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "MongoDB course ObjectId.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "665c9f2e8f1b2a4d7c123456"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Advanced Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Build production-ready applications.",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": "665c9f2e8f1b2a4d7c123457",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Course not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true,
                        "example": "665c9f2e8f1b2a4d7c123456"
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Learn the fundamentals of Laravel.",
                        "nullable": true
                    },
                    "category_id": {
                        "type": "string",
                        "example": "665c9f2e8f1b2a4d7c123457",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}